api/requirements copy.txt
ccxt
websockets
pyyaml
pandas
requests
loguru